# Conflicts:
#	OsmAnd/res/values-bg/strings.xml
#	OsmAnd/res/values-da/strings.xml
#	OsmAnd/res/values-eo/strings.xml
#	OsmAnd/res/values-es-rAR/strings.xml
#	OsmAnd/res/values-es/strings.xml

+++ OsmAnd/src/net/osmand/plus/backup/NetworkSettingsHelper.java
@@ -264,7 +264,7 @@ public class NetworkSettingsHelper extends SettingsHelper {
 			registerSyncBackupTask(key, syncTask);
 			switch (operation) {
 				case SYNC_OPERATION_DELETE -> {
-					if (remoteFile != null) {
+					if (remoteFile != null && !remoteFile.isDeleted()) {
 						if (remoteFile.item != null) {
 							syncTask.deleteItem(remoteFile.item);
 						}
